yc-scraper/requirements.txt
# YC Scraper Dependencies
# Compatible with Python 3.11, 3.12, and 3.13

# Core scraping dependencies
scrapy==2.11.2
selenium==4.15.2
tqdm==4.66.4

# Data analysis (optional - use pandas>=2.2.0 for Python 3.13 support)
pandas>=2.2.0

# Additional helpful packages
lxml==5.1.0  # Required by Scrapy for parsing

# Email generator packages
dnspython
typing
bs4
